page jsf et backing bean transfert

+++ src/main/java/mg/cedric/tpbanque/ejb/GestionnaireCompte.java
@@ -54,7 +54,7 @@ public class GestionnaireCompte {
         return query.getSingleResult();
     }
     
-    public CompteBancaire findById(int id) {  
+    public CompteBancaire findById(long id) {  
         return em.find(CompteBancaire.class, id);  
     }
